rename: CreatePost 위치 변경
